Split models.py into separate Python modules
