+++ apps/openassessment/assessment/models/__init__.py
@@ -1,0 +1,7 @@
+"""
+Export models from each Python module in this package.
+"""
+# pylint: disable=W0401
+
+from .base import *
+from .peer import *
