+++ pylintrc
@@ -161,7 +161,7 @@ bad-names=foo,bar,baz,toto,tutu,tata
 
 # Regular expression which should only match functions or classes name which do
 # not require a docstring
-no-docstring-rgx=__.*__|test_.*|setUp|tearDown
+no-docstring-rgx=__.*__|test_.*|setUp|tearDown|Meta
 
 
 [MISCELLANEOUS]
